requirements.txt

Flask
pytesseract
pandas
openpyxl
Pillow
